Moved fields from the forms file into fields file
